[#60458462] Investigate why twitter stream/lists process dies
- modify twitter related ids to use bigint unsigned
- configure schema dumps to use sql format

+++ app/models/tweet.rb
@@ -9,7 +9,7 @@ class Tweet < ActiveRecord::Base
   searchable do
     text :tweet_text, :stored => true
     time :published_at
-    integer :twitter_profile_id, :multiple => true
+    long :twitter_profile_id, :multiple => true
   end
 
   class << self
